\name{robseg-package}
\alias{robseg-package}
\alias{robseg}
\docType{package}
\title{Segmentation using functionnal pruning for robust losses}
}
\description{A wrapper to a C implementation of robseg}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{robseg}
\packageIndices{robseg}
}
\author{Guillem Rigaill}

Maintainer: \packageMaintainer{robseg}
}

\keyword{ package }


